Fix memory issue involving orb_solver
Hamiltonian is transformed and the energy is checked. This means that
if we are scanning for a good energy, then the original hamiltonian
needs to be stored somewhere.

Before, the original hamiltonian was kept constant by making deepcopy of the hamiltonian
for every energy evaluation.
Now, the hamiltonian is simply rotated back by the same amount.
Hopefully, there are no overflow problems