Allow energy to be fixed in schrodinger equation
In the projected Schrodinger equation (and variants), the energy can be
the computed energy (with respect to some referenc) or be a parameter
that is also optimized. Here, we allow the parameter of the energy to be
fixed such that it is not involved in the optimization process. Then, we
can run the optimization of the energy and of the
wavefunction/hamiltonian parameters separately.