Add starting structure for the parameter mask
In order to allow freezing parameters in the optimization (i.e. remove
them from the objective), the current implementation required a rewrite
of the param_ranges attribute, which admittedly was hard-coded to begin
with. Also, it will be difficult to change the parameter selection after
initialization, meaning that specialized algorithms that utilize
parameter-selecting techniques will be difficult to implment nicely
(well, each step can have an initialization, but that'd be a hacky
solution). This structure will hopefully simplify some of the nastier
parts of the code, particularly the energy-related parts within
least_squares.py

Committed is the proposed structure/class for handling the
selection/update/manipulation of parameters.