Add parent class for parameter containers
Add class that contains attribute `params`, property `nparams` and
method `assign_params`.

Because the Hamiltonian class will also depend on parameters at one
point and it is easier to structure the objective classes. In one of the
objective classes, `SystemEquations`, energy needs to be turned into a
parameter, which was assumed to have the attribute `params`, property
`nparams` and method `assign_params` (just like `BaseWavefunction`).
However, this meant having to wrap around much of these
attribute/property/method into the code. Abstracting out this structure
will be much nicer, I think. Also, it will make thing cleaner when
different kind of parameters are needed, which will happen when we get
complex enough objectives with constraints and whatnot.