Monkey patch property inheritance problem
In LeastSquaresObjective.objective calls the SystemEquations.objective.
In SystemEquations.objective, property num_eqns is needed to set up the
objective values. However, LeastSquaresObjective overwrites this
property to 1, and SystemEquation.objective fails.

Prior to calling SystemEquation.objective, the property num_eqns is
overwritten and then it is changed back after the call.