Add FIXME's
Though it is possible to derivatize wrt the wavefunction and hamiltonian
parameters, there seems to be some problems when we derivatizze wrt the
parameters of the components of the wavefunction (e.g. wavefunctions in
linear combination of wavefunction) and components of hamiltonian (e.g.
perturbed hamiltonian). I don't think that the derivatives can handle
such structures.