Turn method from abstract method to raising NotImplementedError
The method, BaseHamiltonian.assign_integrals, is not always needed to
use the Hamiltonian. In the case of Hamiltonian that is a perturbation
of other Hamiltonians, integrals are supplied by the other Hamiltonians,
so that the method, `assign_integrals` will not be called. In order to
initialize such structures, `assign_integrals` raises
NotImplementedError when called rather than forcing these methods to be
defined.