Add tox file
What:
1. Add separate environment for python distributions, horton, pyscf, keras, and
linters
2. Add flake8-import-order, pep8-naming, black, and bandit to linters

Why:
1. Since horton, pyscf, and keras are optional dependencies, it would be
wasteful wasteful to wait to install these packages for all of the CI tests.
Some of the linters (i.e. black) are specific to a python version, so linters
can be run on a specific python version.
2. Since these linters will replace cardboardlint, it'd be good to have
replacements for import and namespace linters in cardboardlint. The added
linters `flake8-import-order` and `pep8-naming` can be considered as
replacements (in fact, they do a lot more) the `import` and `namespace`,
respectively. Black is a nice auto-formatter that will essentially solve most of
flake8 and pylint complaints. Bandit is probably unnecessary, but doesn't hurt
to have.