Add utility functions for testing
Copied/adapted from https://github.com/theochem/gbasis

What:
1. Add function for initializing a class bypassing its __init__
2. Add function that removes the abstract methods of an abstract class
3. Add function for finding data files
4. Add function for partial derivative via finite difference

Why:
1. Used to test specifically one of the setters
2. Used to test non-abstract methods of an abstract class
3. Even though the find_datafile was removed, it needed to be brought back
because pytest (and other testing framework) executes the tests from outside the
tests directory. Without this function, the data files must be found relative to
the current working directory, turn the tests into a module, or move the data
file to the project module. Since none of these are ideal, the data files are
found relative the the location of the utils.py (which should be in the tests
directory)
4. Though it is not yet used, we can use this code to test all the first order
derivatives.