Remove nose.plugins.attrib.attr
The attr decorator was used to skip the tests that are slow. In pytest, we can
select the tests according to the name of the test. The slow tests have been
labeled by adding "_slow" to the end of the test name. The test can be selected
with `pytest -k "slow" ./tests` and deselected with `pytest -k "not slow"
./tests`