Add exceptions to pylint
1. import-error. Since the package is not installed, none of the imports will
need to be checked. I think there should be better configurations that provide
this functionality without disabling all possible import-error's but this is the
easiest way for me at the moment.
2. len-as-condition. Though it is considered more pythonic to evaluate a list as
a boolean, especially to check if the list is empty, I believe that it is easier
to read if the variable is properly named
3. too-many-locals. There seems to be many intermediate quantities produced
within many of the functions. I think that these variables should be explicitly
defined to ensure that we don't recompute some results and to easily compare
against some equations on paper.