Remove gmpy2 dependency
Seems like it's pointless to keep something for arbitrary precision integer when
the code will reach its limit way before this point. The bitwise operations will
ever be the bottleneck of the code. Since gmpy2 is the only dependency that is
not a standard package for python modules, it is removed. In fact, the
performance difference with just using numpy and python's bitwise operations is
negligible.

What:
1. Replace gmpy2 methods with appropriate python bitwise operation (and some
numpy)
2. Modify API and tests
3. Remove gmpy2 as a dependency in setup and travis and tox