Fix potential bug when computing the nspatial
It seems that the nspatial part was copied from the unrestricted_chemical. Since
one_int is now a numpy array (instead of a tuple of numpy array),
`one_int[0].shape` will get the shape of the first row of the numpy array. This
doesn't actually affect any of the code because one_int is by definition a two
dimensional matrix, so `one_int[0].shape[0]` will still return the same thing.
Nonetheless, the "bug" was fixed